src/requirements.txt
# To load .env
python-dotenv

# Text to ASCII Art
pyfiglet

# Play song of voice
pygame

# Google Text To Speech
gtts

# Speech to Text
speechrecognition

# To connect with AI
openai